+++ src/main/java/next/matcher/MethodMatcher.java
@@ -1,4 +1,4 @@
-package next.hello;
+package next.matcher;
 
 import java.lang.reflect.Method;
