+++ src/test/java/next/hello/HelloCGLibProxyTest.java
@@ -1,5 +1,6 @@
 package next.hello;
 
+import next.interceptor.UppercaseMethodInterceptor;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
