+++ src/test/java/next/hello/JdkProxyTest.java
@@ -1,5 +1,7 @@
 package next.hello;
 
+import next.handler.UpperCaseInvocationHandler;
+import next.interceptor.UppercaseMethodInterceptor;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
